requirements.txt
python-dotenv
loguru
requests
python-telegram-bot
pymongo
pytz